requirements.txt
streamlit
google-generativeai
python-dotenv
langfuse
Pillow
fpdf
langchain-google-genai
langchain-community
faiss-cpu
pypdf